Amend README and add nginx config testing to the runresty script

+++ src/runresty.sh
@@ -16,8 +16,15 @@ do
   TEST_MD5=$(find /src/ -type f -exec md5sum {} \; | md5sum | cut -f1 -d ' ')
   if [ "$FILE_MD5" != "$TEST_MD5" ]
   then
-    echo "Changes detected. Reloading Openresty..."
-    kill -SIGHUP $(pgrep openresty)
-    FILE_MD5=$TEST_MD5
+    # Test nginx config
+    /usr/bin/openresty -c /src/conf/nginx.conf -t
+    if [ $? -eq 0 ]
+    then
+      echo "Changes detected. Reloading Openresty..."
+      kill -SIGHUP $(pgrep openresty)
+      FILE_MD5=$TEST_MD5
+    else
+      echo "Openresty configuration broken, not reloading!"
+    fi
   fi
 done
